requirements.txt
atomicwrites==1.4.0
attrs==21.4.0
beautifulsoup4==4.11.1
bidict==0.22.0
black==22.3.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cloudpickle==2.0.0
colorama==0.4.4
croniter==1.3.4
cycler==0.11.0
dask==2022.4.1
Deprecated==1.2.13
distributed==2022.4.1
dnspython==2.2.1
docker==5.0.3
eventlet==0.33.0
execnet==1.9.0
flaky==3.7.0
Flask==2.1.1
Flask-Cors==3.0.10
Flask-JSON==0.3.4
Flask-SocketIO==5.1.2
fonttools==4.33.3
fsspec==2022.3.0
graphviz==0.20
greenlet==1.1.2
HeapDict==1.0.1
idna==3.3
importlib-metadata==4.11.3
importlib-resources==5.7.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.2
locket==1.0.0
lxml==4.8.0
MarkupSafe==2.1.1
marshmallow==3.15.0
marshmallow-oneofschema==3.0.1
matplotlib==3.5.2
msgpack==1.0.3
mypy==0.812
mypy-extensions==0.4.3
neo4j-driver==4.3.6
neobolt==1.7.17
neomodel==4.0.8
numpy==1.22.3
packaging==21.3
pandas==1.4.2
partd==1.2.0
pathspec==0.9.0
pendulum==2.1.2
Pillow==9.1.0
platformdirs==2.5.2
plotly==5.7.0
pluggy==1.0.0
prefect==1.2.0
psutil==5.9.0
py==1.11.0
pycparser==2.21
PyGithub==1.55
Pygments==2.11.2
PyJWT==2.3.0
PyNaCl==1.5.0
pyparsing==3.0.8
pytest==7.1.1
pytest-env==0.6.2
pytest-forked==1.4.0
pytest-xdist==2.5.0
python-box==6.0.2
python-dateutil==2.8.2
python-dotenv==0.20.0
python-engineio==4.3.2
python-slugify==6.1.1
python-socketio==5.6.0
pytz==2022.1
pytzdata==2020.1
pywin32==227
PyYAML==6.0
requests==2.27.1
responses==0.20.0
Shapely==1.7.1
six==1.16.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tabulate==0.8.9
tblib==1.7.0
tenacity==8.0.1
testfixtures==6.18.5
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
toolz==0.11.2
tornado==6.1
typed-ast==1.4.3
typing_extensions==4.2.0
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.1
wrapt==1.14.0
zict==2.1.0
zipp==3.8.0
